vault-cluster/api/requirements.txt
#flask==1.1.2
#boto==2.49.0
#boto3==1.14.59
#aws==0.2.5
#flask_restful==0.3.8
#jsonify==0.5

flask
boto
boto3
aws
flask_restful
jsonify
python-dotenv
hvac
certifi
chardet
idna
pyhcl
requests
six
urllib3
Django
djangorestframework
pytz
sqlparse
dj-database-url
gunicorn
psycopg2-binary